refactor(chart-advisor): grouped auto-chart
