+++ packages/chart-advisor/src/auto-chart/constants.ts
