+++ packages/chart-advisor/src/auto-chart/toolbar.ts
@@ -1,8 +1,8 @@
 import Thumbnails from '@antv/thumbnails';
 import { CKBJson, ChartID } from '@antv/knowledge';
-import { CLASS_PREFIX } from './style';
-import { Advice, adviceToLibConfig } from './advice-pipeline';
-import { getElementDisplay } from './util';
+import { CLASS_PREFIX } from '../style';
+import { Advice, adviceToLibConfig } from '../advice-pipeline';
+import { getElementDisplay } from '../util';
 import { AutoPlot } from './auto-plot';
 import { customChartType } from './custom-plot';
 import { intl, getLanguage } from './i18n';
